requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
openpyxl==3.1.2
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
scikit-learn==1.4.0
scipy==1.12.0
matplotlib==3.8.2
reportlab==4.0.9
python-docx==1.1.0
